pad memory areas to survive MUL w overreaching offset
Consider a program like '[-<<<+>>>]'. Since cells are initialized to
0, the loop will not be entered and the program will exit
successfully. However, with awib's mutliplication loop optimization,
this program will be compiled into LMUL(3, 1). A naive backend would
compile this into an unconditional write to index -3, resulting in
undefined behaviour even though the program is correct.

To handle this, all backends now have padded memory areas. The max
offset for LMUL and RMUL is 127, so twice that many cells have been
added and the pointer is now initialized to p=127.